MakeFile done. Errors corrected
