+++ .gitignore
@@ -1,2 +1,3 @@
 *.pdf
 test.c
+*.o
