+++ utils.c
@@ -1,4 +1,4 @@
-#include <stdio.h>
+#include "utils.h"
 
 void print_help(void){
     printf("-a , --all \t\t- a informação exibida diz respeito também a ficheiros\n");
@@ -10,16 +10,6 @@ void print_help(void){
     printf("--max-depth=N \t\t- limita a informação exibida a N (0,1, ...) níveis de profundidade de diretorios\n ");
 }
 
-void transformToString(char *result, char *array[], int n){
-    int i=0;
-    while(array[i] != NULL && i<n){
-        if (i==0) 
-            sprintf(result, "%s", array[i]);
-        else
-            sprintf(result, "%s %s", result, array[i]);
-        i++;
-    }
-}
 
 void copy_values(char *dest[], char *copy[], int n){
     for (int i =0; i<n; i++){
