+++ utils.h
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+
+
+void print_help(void);
+
+void copy_values(char *dest[], char *copy[], int n);
